# doc/api.yaml
openapi: 3.0.3
...
info:
  title: WASAPhoto
  description: Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can
              upload your photos directly from your PC, and they will be visible to everyone following you.
  version: 1.0.0
paths:
  /session:
    post:
      tags: ["Login"]
      summary: Logs in the user
      description: |-
        If the user does not exists, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              type: object
              description: User login details.
              properties:
                name:
                  type: string
                  example: Maria
                  pattern: '^.*?$'
                  description: "Username of the user who is logging in"
                  minLength: 3
                  maxLength: 16
        required: true
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                description: User identifier
                properties:
                  identifier:
                    type: string
                    description: The user's identifier
                    example: "abcdef012345"
  /photos:
    post:
      tags: ["Photos"]
      summary: Upload a new picture
      operationId: uploadPhoto
      description: Upload a new picture
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Photo successfully added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    get:
      tags: ["Photos"]
      operationId: getPhotos
      summary: Get all photos
      description: Get all photos
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Photos Retrieved
          content:
            application/json:
              schema:
                type: object
                description: List of photos
                properties:
                  photos:
                    type: array
                    description: List of photos
                    minItems: 1
                    maxItems: 100
                    items:
                      $ref: "#/components/schemas/Photo"
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /photos/{pid}:
    parameters:
      - $ref: "#/components/parameters/photo_id"
    get:
      tags: ["Photos"]
      operationId: getPhoto
      summary: Get a Photo
      description: Get a Photo
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Photo Retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    delete:
      tags: ["Photos"]
      operationId: deletePhoto
      summary: Delete a Photo
      description: Delete a Photo
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Photo Deleted
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /users/{uid}/profile:
    parameters:
      - $ref: "#/components/parameters/user_id"
    get:
      operationId: getUserProfile
      tags: ["Users"]
      summary: Get User Profile
      description: Get User Profile
      security:
        - bearerAuth: []
      responses: 
        "200":
          description: User Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        "404": { $ref: "#/components/responses/NotFoundError" }
  /user/{uid}/username:
    parameters:
      - $ref: "#/components/parameters/user_id"
    put:
      tags: ["Users"]
      operationId: setMyUserName
      summary: Update my username
      description: Allows to set the username
      security:
        - bearerAuth: []
      requestBody:
        description: JSON Object with my username
        required: true
        content:
          application/json:
            schema:
              type: object
              description: User's new username
              properties: 
                username: 
                  type: string
                  pattern: '^[a-zA-Z0-9-_]+$'
                  minLength: 3
                  maxLength: 12
                  description: "The new username"
                  example: "new_username"
      responses:
        "200": 
          description: Username updated sucessfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /user/{uid}/stream:
    parameters:
      - $ref: "#/components/parameters/user_id"
    get:
      operationId: getMyStream
      tags: ["Users"]
      summary: Get User Stream
      description: Get User Stream
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User Stream
          content:
            application/json:
              schema:
                type: object
                description: List of photos
                properties:
                  stream:
                    type: array
                    description: List of photos
                    minItems: 1
                    maxItems: 100
                    items:
                      $ref: "#/components/schemas/Photo"
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
components:
  schemas:
    Photo:
      type: object
      description: Photo
      properties:
        photo_id: 
          type: string
          description: The photo's identifier
          example: "abcdef012345"
        user_id:
          type: string
          description: The user's identifier
        url:
          type: string
          description: The photo's URL
        date:
          type: string
          description: The photo's date
          format: date-time
      required:
        - photo_id
    User:
      type: object
      description: User
      properties:
        user_id:
          type: string
          description: The user's identifier
        username:
          type: string
          description: The user's username
      required:
        - user_id
    Profile:
      type: object
      description: User Profile
      properties:
        user_id:
          type: string
          description: The user's identifier
        username:
          type: string
          description: The user's username
        followers:
          type: integer
          description: Number of followers
          example: 0
          minimum: 0
          maximum: 9999
        followees:
          type: integer
          description: Number of followees
          example: 0
          minimum: 0
          maximum: 9999
        photo_count:
          type: integer
          description: Number of photos uploaded by the user
          example: 0
          minimum: 0
          maximum: 9999
...
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc)
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs
    UnauthorizedError:
      description: Access token is missing or invalid
    NotFoundError:
      description: The requested resource was not found
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    photo_id:
        description: The photo's identifier
        schema:
          type: string
          pattern: '^[a-zA-Z0-9-_]+$'
          minLength: 12
          maxLength: 12
        name: pid
        in: path
        required: true
        allowEmptyValue: false
    user_id:
      description: The user's identifier
      schema:
        type: string
        pattern: '^[a-zA-Z0-9-_]+$'
        minLength: 12
        maxLength: 12
      name: uid
      in: path
      required: true
      allowEmptyValue: false
...